refactor: Router 로직 폴더로 분리
